+++ src/common/PhysicsLX56_Projectiles.cpp
@@ -548,7 +548,7 @@ float Proj_SpawnParent::fixedRandomFloat() const {
 CVec Proj_SpawnParent::position() const {
 	switch(type) {
 		case PSPT_NOTHING: return CVec(0,0);
-		case PSPT_SHOT: return shot->cPos + GetVecFromAngle(shot->nAngle) * 8;
+		case PSPT_SHOT: return shot->cPos + GetVecFromAngle((float)shot->nAngle) * 8;
 		case PSPT_PROJ: return proj->getPos();
 	}
 	return CVec(0,0);
